version 1.3.1-beta1

+++ src/InstagramCSharp/Models/Tag.cs
@@ -1,9 +1,8 @@
-﻿
-namespace InstagramCSharp.Models
+﻿namespace InstagramCSharp.Models
 {
     public class Tag
     {
-        public int Media_count { get; set; }
+        public long Media_count { get; set; }
         public string Name { get; set; }
     }
 }
